refactor(Validator): 검증 추상화
