+++ src/main/java/util/validator/CarNameValidator.java
@@ -1,21 +1,13 @@
-package util;
+package util.validator;
 
-import java.util.List;
+public class CarNameValidator extends Validator {
+    private static final int MAX_NAME_LENGTH = 5;
 
-public class CarNameValidator {
-    public static final int MIN_CAR_NUMBER = 2;
-    public static final int MAX_NAME_LENGTH = 5;
-
-    public void validate(List<String> names) {
-        validateCarNumber(names);
-        names.forEach(this::validateCarNameLength);
+    @Override
+    public void validate(String name) {
+        validateCarNameLength(name);
     }
 
-    private void validateCarNumber(List<String> names) {
-        if (names.size() < MIN_CAR_NUMBER) {
-            throw new IllegalArgumentException(Message.EXCEPTION_CAR_NUMBER.message);
-        }
-    }
 
     private void validateCarNameLength(String name) {
         if (name.isBlank()) {
@@ -27,7 +19,6 @@ public class CarNameValidator {
     }
 
     public enum Message {
-        EXCEPTION_CAR_NUMBER("2개 이상의 자동차를 입력해 주세요."),
         EXCEPTION_CAR_NAME_LENGTH("5자 이하의 이름을 입력해주세요."),
         EXCEPTION_CAR_NAME_BLANK("공백은 이름으로 사용할 수 없습니다");
         private final String message;
