+++ src/main/java/util/validator/Validator.java
@@ -1,0 +1,7 @@
+package util.validator;
+
+public abstract class Validator {
+
+    public abstract void validate(String input);
+
+}
