+++ src/test/java/util/MoveCountValidatorTest.java
@@ -9,7 +9,8 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import util.MoveCountValidator.Message;
+import util.validator.MoveCountValidator;
+import util.validator.MoveCountValidator.Message;
 
 class MoveCountValidatorTest {
     private MoveCountValidator moveCountValidator;
